moved knex file into config directory
