+++ packages/backend/config/database.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const env = process.env.NODE_ENV || 'development'
-const knexfile = require('../knexfile')
+const knexfile = require('./knexfile')
 const knex = require('knex')(knexfile[env])
 
 module.exports = knex
